+++ src/main/java/com/goormdari/domain/routine/presentation/RoutineController.java
@@ -1,12 +1,11 @@
 package com.goormdari.domain.routine.presentation;
 
 
-import com.goormdari.domain.calendar.dto.response.CheckGoalProgressResponse;
 import com.goormdari.domain.calendar.exception.InvalidTokenException;
 import com.goormdari.domain.routine.application.RoutineService;
 import com.goormdari.domain.routine.domain.Routine;
 import com.goormdari.domain.routine.dto.request.CompleteRoutineRequest;
-import com.goormdari.domain.user.domain.service.UserService;
+import com.goormdari.domain.user.service.UserService;
 import com.goormdari.global.config.security.jwt.JWTUtil;
 import com.goormdari.global.config.s3.S3Service;
 import com.goormdari.global.payload.ErrorResponse;
@@ -97,4 +96,9 @@ public class RoutineController {
         return ResponseCustom.OK(routineService.findAllRoutineByUserId(userId));
     }
 
+    @PostMapping("/temp/uploader")
+    public String uploadTempImage(@ModelAttribute MultipartFile file) {
+        String url = s3Service.uploadImageToS3(file);
+        return url;
+    }
 }
