--- src/main/java/com/goormdari/domain/user/domain/dto/response/findCurrentStepResponse.java
@@ -1,7 +1,0 @@
-package com.goormdari.domain.user.domain.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record findCurrentStepResponse (int currentStep) {
-}
