+++ src/main/java/com/goormdari/domain/user/dto/request/AddUserRequest.java
@@ -1,4 +1,4 @@
-package com.goormdari.domain.user.domain.dto.request;
+package com.goormdari.domain.user.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
