+++ src/main/java/com/goormdari/domain/user/dto/response/FindCurrentStepResponse.java
@@ -1,0 +1,7 @@
+package com.goormdari.domain.user.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record FindCurrentStepResponse(int currentStep) {
+}
