+++ src/main/java/com/goormdari/domain/user/presentation/AuthController.java
@@ -1,9 +1,9 @@
 package com.goormdari.domain.user.presentation;
 
-import com.goormdari.domain.user.domain.dto.request.AddUserRequest;
-import com.goormdari.domain.user.domain.dto.response.JwtResponse;
-import com.goormdari.domain.user.domain.dto.request.LoginRequest;
-import com.goormdari.domain.user.domain.service.UserService;
+import com.goormdari.domain.user.dto.request.AddUserRequest;
+import com.goormdari.domain.user.dto.response.JwtResponse;
+import com.goormdari.domain.user.dto.request.LoginRequest;
+import com.goormdari.domain.user.service.UserService;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
