+++ src/main/java/com/goormdari/domain/user/presentation/UserController.java
@@ -1,9 +1,9 @@
 package com.goormdari.domain.user.presentation;
 
 import com.goormdari.domain.calendar.exception.InvalidTokenException;
-import com.goormdari.domain.user.domain.dto.response.UserInfoResponse;
-import com.goormdari.domain.user.domain.service.UserService;
-import com.goormdari.domain.user.domain.dto.response.findCurrentStepResponse;
+import com.goormdari.domain.user.dto.response.UserInfoResponse;
+import com.goormdari.domain.user.service.UserService;
+import com.goormdari.domain.user.dto.response.FindCurrentStepResponse;
 import com.goormdari.global.config.security.jwt.JWTUtil;
 import com.goormdari.global.payload.ErrorResponse;
 import com.goormdari.global.payload.ResponseCustom;
@@ -29,11 +29,11 @@ public class UserController {
 
     @Operation(summary = "현재 루틴 수 조회", description = "사용자가 완수한 루틴의 개수")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "조회 성공 ", content = {@Content(mediaType = "application/json", schema = @Schema(implementation =  findCurrentStepResponse.class))}),
+            @ApiResponse(responseCode = "200", description = "조회 성공 ", content = {@Content(mediaType = "application/json", schema = @Schema(implementation =  FindCurrentStepResponse.class))}),
             @ApiResponse(responseCode = "400", description = "조회 실패", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))}),
     })
     @GetMapping("/current-step")
-    public ResponseCustom<findCurrentStepResponse> getCurrentStep(
+    public ResponseCustom<FindCurrentStepResponse> getCurrentStep(
             @Parameter(description = "Accesstoken을 입력해주세요.", required = true) @RequestHeader("Authorization") String token
     ) {
         if (token == null) {
