+++ src/main/java/com/goormdari/domain/user/service/UserService.java
@@ -1,14 +1,13 @@
-package com.goormdari.domain.user.domain.service;
+package com.goormdari.domain.user.service;
 
 import com.amazonaws.services.kms.model.NotFoundException;
-import com.goormdari.domain.user.domain.dto.response.UserInfoResponse;
-import com.goormdari.domain.user.domain.dto.response.findCurrentStepResponse;
+import com.goormdari.domain.user.dto.response.UserInfoResponse;
+import com.goormdari.domain.user.dto.response.FindCurrentStepResponse;
 import com.goormdari.domain.user.domain.User;
 import com.goormdari.domain.user.domain.DefaultProfileUrl;
-import com.goormdari.domain.user.domain.dto.request.AddUserRequest;
-import com.goormdari.domain.user.domain.dto.request.LoginRequest;
-import com.goormdari.domain.user.domain.dto.response.JwtResponse;
-import com.goormdari.domain.user.domain.dto.response.findCurrentStepResponse;
+import com.goormdari.domain.user.dto.request.AddUserRequest;
+import com.goormdari.domain.user.dto.request.LoginRequest;
+import com.goormdari.domain.user.dto.response.JwtResponse;
 import com.goormdari.domain.user.domain.repository.UserRepository;
 import com.goormdari.global.config.security.jwt.JWTUtil;
 
@@ -35,11 +34,11 @@ public class UserService {
     private final JWTUtil jwtUtil;
 
     @Transactional
-    public findCurrentStepResponse findCurrentStepById(Long userId) {
+    public FindCurrentStepResponse findCurrentStepById(Long userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(()->new NotFoundException("User Not Found"));
 
-        return findCurrentStepResponse.builder().currentStep(user.getCurrentStep()).build();
+        return FindCurrentStepResponse.builder().currentStep(user.getCurrentStep()).build();
     }
 
     @Transactional
